Remove publishing from feature branches
We received a complaint that we publish too frequently to maven central. I suspect this is due to the fact that excavators like to push commits frequently. In order to be good internet citizens, we should publish to maven local on our trial branches and only exercise full publishing on tags.